split up string literal rules
