+++ src/index.ts
@@ -1,5 +1,8 @@
 export {Rule as SwitchIdentifiersRule} from './rules/switchIdentifiersRule';
-export {Rule as SwitchStringLiteralsRule} from './rules/switchStringLiteralsRule';
+export {Rule as SwitchStringLiteralsRule} from './rules/switchStringLiteralAttributeSelectorsRule';
 export {Rule as SwitchTemplatesRule} from './rules/switchTemplatesRule';
 export {Rule as SwitchStylesheetsRule} from './rules/switchStylesheetsRule';
 export {Rule as SwitchPropertyNamesRule} from './rules/switchPropertyNamesRule';
+export {Rule as SwitchStringLiteralAttributeSelectorsRule} from './rules/switchStringLiteralAttributeSelectorsRule';
+export {Rule as SwitchStringLiteralCssNamesRule} from './rules/switchStringLiteralCssNamesRule';
+export {Rule as SwitchStringLiteralElementSelectorsRule} from './rules/switchStringLiteralElementSelectorsRule';
